requirements.txt
attrs==22.2.0
Automat==22.10.0
certifi==2022.9.24
click==8.0.4
constantly==15.1.0
dataclasses==0.8
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Twisted==0.1.2
greenlet==2.0.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.8.3
incremental==22.10.0
influxdb-client==1.31.0
itsdangerous==2.0.1
Jinja2==3.0.3
kafka-python==2.0.2
MarkupSafe==2.0.1
mysql-connector==2.2.9
observable==1.0.3
pip==9.0.3
python-dateutil==2.8.2
Rx==3.2.0
setuptools==39.2.0
six==1.16.0
SQLAlchemy==1.4.34
style==1.1.0
Twisted==22.4.0
typing-extensions==4.1.1
update==0.0.1
urllib3==1.26.12
Werkzeug==2.0.3
zipp==3.6.0
zope.interface==5.5.2
xmltodict==0.13.0
requests==2.27.1